sbt-ci-build-doc.bsh and docker-version-docs.bsh script
This is a general purpose refactor of sbt-ghpages-ssh.sh to support building
and pushing pages to alternate targets (i.e. when run on internal CI systems).

`DOCS_NOPUBLISH` flag to control publication
`FORCE_DOCS` flag forces documentation to be built even when CI environment would
prevent it (i.e. locally, or on the wrong branch).

GH pages support for the dockermake.bsh script.

This is implemented via two changes: first, the dockermake script now outputs
a list of pushed images in line-delimited format to a file called .pushed_images

Secondly, to build and update the GH pages version documentation the script
docker-version-docs.bsh must be called with the list of pages as input.

In practice this means calling ./docker-version-docs.bsh $(cat .pushed_images)
for dockermake, but it allows standalone images to use other systems.