sbt-ci-build-doc.bsh
This is a general purpose refactor of sbt-ghpages-ssh.sh to support building
and pushing pages to alternate targets (i.e. when run on internal CI systems).

`DOCS_NOPUBLISH` flag to control publication
`FORCE_DOCS` flag forces documentation to be built even when CI environment would
prevent it (i.e. locally, or on the wrong branch).